refactor: keep more utils internal
