+++ packages/common/src/search/index.ts
@@ -10,5 +10,4 @@ export * from "./types";
 export * from "./collectionState";
 export * from "./convertCatalog";
 export * from "./filter-utils";
-export * from "./ifilter-utils";
 export * from "./hubSearch";
